트리 level 버그 수정(포스트플랍 첫 벳 팟1종 붕괴 = 번역 16-27% 진범) + 설계 C 확정(첫 벳 8종/리레이즈 팟만, 정보집합 12M) + float32 — 재솔빙 2000반복

+++ poker-cfrplus/hunl_solver.py
@@ -63,8 +63,8 @@ class HunlCFRPlus:
                 kb = PRE if n.street == 0 else K
                 acts = sorted(n.children)
                 n.acts = acts
-                n.regret = np.zeros((kb, len(acts)))
-                n.ssum = np.zeros((kb, len(acts)))
+                n.regret = np.zeros((kb, len(acts)), dtype=np.float32)
+                n.ssum = np.zeros((kb, len(acts)), dtype=np.float32)
                 self._nodes.append(n)
                 for c in n.children.values():
                     attach(c)
